Add sorting and fix masonry layout
I used https://github.com/jessekorzan/masonry-css-js/blob/master/src/App.js as reference to reorder the articles array in order for them to be display from left to right in the CSS columns version of a masonry grid. Otherwise, the articles are listed top to bottom first, and then left to right which is a no-no.